Use global variable for identifying postcode codes json file. Add in semi-colon to avoid merged JS errors
We cannot assume that there is always a JS directory at the root as Magento may be installed in a sub-directory.
We would normally pass in the base URL as a class parameter but over refactoring this entire file I've declared a global
variable in the template which can then be used by this file.

Without the semi-colon before the start of the function, there's an error upon merging of JS which breaks the rule-builder.